+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/response/CreateAuthMethodResponse.java
@@ -36,7 +36,7 @@ public class CreateAuthMethodResponse {
     private Integer userPrefsColumn;
     private Boolean userPrefsDefault;
     private Boolean checkAuthFails;
-    private Long maxAuthFails;
+    private Integer maxAuthFails;
     private Boolean hasUserInterface;
     private Boolean hasMobileToken;
     private String displayNameKey;
