+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/response/CreateCredentialPolicyResponse.java
@@ -38,9 +38,9 @@ public class CreateCredentialPolicyResponse {
     private Integer credentialLengthMin;
     private Integer credentialLengthMax;
     private String credentialAllowedChars;
-    private Long limitSoft;
-    private Long limitHard;
-    private long checkHistoryCount;
+    private Integer limitSoft;
+    private Integer limitHard;
+    private int checkHistoryCount;
     private boolean rotationEnabled;
     private Integer rotationDays;
     private String usernameGenAlgorithm;
