+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/response/UpdateCredentialPolicyResponse.java
@@ -37,8 +37,8 @@ public class UpdateCredentialPolicyResponse {
     private Integer credentialLengthMin;
     private Integer credentialLengthMax;
     private String credentialAllowedChars;
-    private Long limitSoft;
-    private Long limitHard;
+    private Integer limitSoft;
+    private Integer limitHard;
     private int checkHistoryCount;
     private boolean rotationEnabled;
     private Integer rotationDays;
